- looks I missed a few shorts.
Now all local short variables are gone, except for 2 lotags arrays.

+++ source/games/duke/src/player_d.cpp
@@ -2704,7 +2704,7 @@ void processinput_d(int snum)
 	int j, k, doubvel, fz, cz, truefdist;
 	Collision chz, clz;
 	bool shrunk;
-	short psect, psectlotag;
+	int psect, psectlotag;
 	struct player_struct* p;
 	spritetype* s;
